pyproject.toml
[project]
name = "genai-insight-agent"
version = "0.1.0"
description = "GenAI项目洞察分析代理"
authors = [{name = "ybalbert001", email = "ybalbert001@gmail.com"}]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "schedule>=1.2.0",
    "requests>=2.31.0",
    "json-repair>=0.25.0",
    "python-dotenv>=1.0.0",
    "jinja2>=3.1.0",
    "matplotlib>=3.8.0",
    "markdown2>=2.4.0",
    "premailer>=3.10.0",
    "pyyaml>=6.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = []
